+++ src/LuaScriptBuilder.php
@@ -237,7 +237,7 @@ EOF;
             return $str;
         }
 
-        return "'" . addcslashes($str, "\000\n\r\\\032\047") . "'";
+        return "'" . addcslashes($str, "\000\n\r\\\047") . "'";
     }
 
     /**